requirements.txt
langchain
langchain-openai
langchain-community
langgraph
langchain-upstage
tavily-python
openai
faiss-cpu
chromadb
pypdf
pypdf2
python-dotenv
pydantic
jupyter
ipykernel
matplotlib
graphviz
pandas

